Update README.md
Updated:
Conditionals based on Time
Conditionals based on Camera Status

Included the new '==' equals condition for both time and camera status, and added in camera status system variable 'v' along with the applicable values and two examples. Added an example for time as well, based on the same format as the other examples.